Refactor note scheduling and music player modules
Extracted NoteScheduler, MusicPlayer, and LanguageWindow into separate modules for better maintainability. Moved system info logging to ConfigManager. Improved pause key handling and playback state logging. Updated imports and removed redundant code from ProjectLyrica.py. Minor logging and error handling improvements throughout.
